requirements.txt
transformers
pillow
streamlit
requests
torch